requirements.txt
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
darkdetect==0.8.0
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
faiss==1.7.4
fastjsonschema==2.19.1
filelock==3.9.0
fsspec==2023.4.0
h5py==3.10.0
idna==3.7
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
networkx==3.2.1
numpy @ file:///D:/bld/numpy_1707225528937/work/dist/numpy-1.26.4-cp311-cp311-win_amd64.whl#sha256=537f0573fefb1e93b3633e8a79102e6b6d2911e09e0a90008993efef4242448e
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==5.26.1
pure-eval==0.2.2
Pygments==2.17.2
PyQt6==6.6.1
PyQt6-Fluent-Widgets==1.5.4
PyQt6-Frameless-Window==0.3.9
PyQt6-Qt6==6.6.3
PyQt6-sip==13.6.0
python-dateutil==2.9.0.post0
pywin32==306
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tensorboardX==2.6.2.2
threadpoolctl==3.4.0
tinycss2==1.2.1
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchvision==0.17.2+cu121
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.8.0
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
